♻️ extract cell renderer
